feat(led): make it optional

+++ src/Models/RadioModels.h
@@ -82,7 +82,7 @@ namespace Victoria {
   };
 
   struct RadioModel {
-    int inputPin = -1;
+    int8_t inputPin = -1;
     std::vector<RadioRule> rules;
     std::vector<RadioCommand> commands;
   };
